pyproject.toml
[tool.poetry]
name = "bookstore-02"
version = "0.1.0"
description = ""
authors = ["Anderson Costa <costaandersom23@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.13"
django = ">=5.2.1,<6.0.0"
djangorestframework = ">=3.16.0,<4.0.0"
django-extensions = "^4.1"
requests = "^2.32.4"
python-slugify = "^8.0.4"
django-debug-toolbar = "^5.2.0"
python-dotenv = "^1.1.0"
psycopg2-binary = "^2.9.10"
django-cors-headers = "^4.7.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.0"
pytest-django = "^4.11.1"
black = "^25.1.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
